Add ability to restart the FTP client while retaining the FUSE host
Adds the method `SetAddress` to the FTP client which enables it to
quit the current connections and restart using another address. This
is primarily intended for use cases where the FUSE mount must survive
a change of remote FTP address.

Adds a `dir` parameter is also added to the `NewFTPClient` function.
The client will tell the server to change to this directory after each
successful connect.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>