Handle mismatch in offset during write
The offset passed in the write method isn't always aligned with the
offset of the previous write call to the same file descriptor. In
such situations, the current store operation must be closed and allowed
to drain, before a new store operation starts at the new offset.

This commit ensures correct operation of the write offset and separates
the write and read offsets.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>